clearify ejson_parser
